refactor: directory 추가
1. views, models, controllers 추가
