+++ cypress/e2e/lotto.cy.js
@@ -1,5 +1,5 @@
 import { ERROR_MESSAGE } from '../../src/const';
-import Lotto from '../../src/js/model/lotto';
+import Lotto from '../../src/js/models/lotto';
 
 const lotto = new Lotto();
