+++ src/js/app.js
@@ -1,11 +1,9 @@
-import InputPriceForm from './view/input-price-form.js';
-import PurchasedLotto from './view/purchased-lotto.js';
-import WinningInputForm from './view/winning-input-form.js';
+import InputPriceForm from './views/input-price-form.js';
+import PurchasedLotto from './views/purchased-lotto.js';
+import WinningInputForm from './views/winning-input-form.js';
 
 class App {
   constructor($target, lotto) {
-    this.lotto = lotto;
-    this.$target = $target;
     new InputPriceForm($target, lotto);
     new PurchasedLotto($target, lotto);
     new WinningInputForm($target, lotto);
