+++ src/js/index.js
@@ -1,4 +1,4 @@
 import App from './app.js';
-import Lotto from './model/lotto.js';
+import Lotto from './models/lotto.js';
 
 new App(document.querySelector('#app'), new Lotto());
