+++ src/js/views/input-price-form.js
@@ -1,5 +1,5 @@
 import { $ } from '../../utils.js';
-import Controller from '../controller.js';
+import Controller from '../controllers/controller.js';
 
 class InputPriceForm extends Controller {
   constructor($target, lotto) {
