+++ src/js/views/purchased-lotto.js
@@ -1,5 +1,5 @@
 import { $ } from '../../utils.js';
-import Controller from '../controller.js';
+import Controller from '../controllers/controller.js';
 
 class PurchasedLotto extends Controller {
   constructor($target, lotto) {
